refac(ton): move ton to coin-framework
