+++ libs/coin-modules/coin-ton/src/errors.ts
@@ -1,0 +1,6 @@
+import { createCustomErrorClass } from "@ledgerhq/errors";
+
+/*
+ * When the recipient is a new named account, and needs to be created first.
+ */
+export const TonCommentInvalid: any = createCustomErrorClass("TonCommentInvalid");
