+++ libs/coin-modules/coin-ton/src/signer.ts
@@ -1,0 +1,9 @@
+import { Cell } from "@ton/ton";
+import { TonTransport } from "@ton-community/ton-ledger";
+
+export type TonAddress = {
+  publicKey: Buffer;
+  address: string;
+};
+export type TonSignature = Cell | undefined;
+export interface TonSigner extends TonTransport {};
