+++ libs/ledger-live-common/package.json
@@ -137,6 +137,7 @@
     "@ledgerhq/coin-evm": "workspace:^",
     "@ledgerhq/coin-framework": "workspace:^",
     "@ledgerhq/coin-near": "workspace:^",
+    "@ledgerhq/coin-ton": "workspace:^",
     "@ledgerhq/coin-polkadot": "workspace:^",
     "@ledgerhq/compressjs": "github:LedgerHQ/compressjs#d9e8e4d994923e0ea76a32b97289bcccfe71b82e",
     "@ledgerhq/crypto-icons-ui": "workspace:^",
@@ -188,8 +189,6 @@
     "@taquito/utils": "^13.0.1",
     "@ton-community/ton-ledger": "^7.0.1",
     "@ton/core": "^0.56.1",
-    "@ton/crypto": "^3.2.0",
-    "@ton/ton": "^13.11.1",
     "@types/bchaddrjs": "^0.4.0",
     "@types/bs58check": "^2.1.0",
     "@types/pako": "^2.0.0",
