--- libs/ledger-live-common/src/families/ton/bridge/js.ts
@@ -1,7 +1,0 @@
-import { accountBridge } from "./account";
-import { currencyBridge } from "./currency";
-
-export default {
-  currencyBridge,
-  accountBridge,
-};
