--- libs/ledger-live-common/src/families/ton/errors.ts
@@ -1,3 +1,0 @@
-import { createCustomErrorClass } from "@ledgerhq/errors";
-
-export const TonCommentInvalid = createCustomErrorClass("TonCommentInvalid");
